Feat: make model and repository that auth service

+++ jumunseo/lib/features/login/model/token_model.dart
@@ -1,5 +1,3 @@
-import 'dart:ffi';
-
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'token_model.g.dart';
@@ -8,7 +6,7 @@ part 'token_model.freezed.dart';
 @freezed
 class TokenModel with _$TokenModel{
   factory TokenModel(
-      {required Int64 expiredTime,
+      {required int expiredTime,
       required String accessToken}) = _TokenModel;
 
   factory TokenModel.fromJson(Map<String, dynamic> json) =>
